feat: style control divs to fit pwrd rqmts
Moved margin from bottom to top; this allows for appropriate spacing
between the input control divs, without adding margin to the bottom
of the password divs.

Added min-height to password divs using trial and error. This gives a
temporary height to the divs that can be calculated more precisely
later. Having the height keeps the divs from resizing when the
password requirements are added via JS.